Rename lib to components
