+++ terminal-demo-app/src/components/Group/Group.jsx
