+++ terminal-demo-app/src/components/TextInput/TextInput.jsx
